put in /usr/share/gnome-shell/js/ui/status
Added mousetweaks to the gnome accessibility menu and now it is on by default the user must turn it off at boot if they don't want it. 

This potentially solves an issue for those who cannot use a mouse click at all due to disability and otherwise would not able to turn on the hover to click feature in order to use it. 

It has a nice side effect that mousetweaks is much easier to quickly disable once running now too (before you had to kill it from the command-line or from a task manager - or at least that is the only way I had found) 

The downside of this is that it might introduce an annoyance for people who do not need mousetweaks because they may have to turn it off every time they login. (Need to check this properly)

==========
TODO 

make it easy to permanently disable for those who don't wish for it.

find a decent icon from somewhere

put some comments in this file to explain how it works.